pyproject.toml
[project]
name = "nonebot-plugin-telegrime"
version = "0.1.0"
description = ""
authors = [
    { name = "Jigsaw", email = "j1g5aw@foxmail.com" },
]
dependencies = [
    "nonebot2<3.0.0,>=2.0.0rc2",
    "nonebot-adapter-telegram @ git+ssh://github.com/nonebot/adapter-telegram",
]
requires-python = ">=3.10,<4.0"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
pre-commit = ["pycln", "isort", "black", "nonemoji", "pre-commit"]
nonebot = [
    "httpx>=0.23.3",
]

[tool.pdm.scripts]
post_install = "pre-commit install"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.pycln]
path = "."
all = false

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
